backend/requirements.txt
python-socketio[asyncio]==5.11.0
aiohttp==3.9.1
psycopg2-binary>=2.9.0

